Extract secondary utility panes
